+++ mm2src/coins/tendermint/iris/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod ethermint_account;
 pub(crate) mod htlc;
 pub(crate) mod htlc_proto;